Fix tests
- ProfileHeaderView tests has turn into UsernameDetailView tests
- IncomingConnectionView was also using usernameDetailView and have been fixed
