+++ Wire-iOS/Sources/UserInterface/UserProfile/Views/UserNameDetailView.swift
@@ -1,4 +1,4 @@
-//
+ //
 // Wire
 // Copyright (C) 2016 Wire Swiss GmbH
 //
